requirements.txt
ai4bharat-transliteration>=1.1.3
flask>=2.0.0
requests>=2.25.0 